Python code split to separate file. File is doxygenized now.
