+++ sw/psg14a_emu/.idea/psg14a_emu.iml
@@ -1,0 +1,8 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<module version="4">
+  <component name="FacetManager">
+    <facet type="Python" name="Python facet">
+      <configuration sdkName="Python 3.10" />
+    </facet>
+  </component>
+</module>
