+++ sw/psg14a_emu/sources/bl/src/control_tables.c
@@ -7,103 +7,9 @@
 
 #include "control_tables.h"
 /*[[[cog
-import cog
-import re
-def checkCyclogram(split):
-  return len(split) >= 3 and split[0] == 'const' and split[1] == 'Channel_Step_t' and split[2].startswith('Cyclogram_')
+from cycle2uml import convert_to_uml
 
-def line2dict(line):
-  ret = {}
-  line_split = re.findall("\.[A-z0-9]+\s*=\s*[A-z0-9]+", line)
-  for item in line_split:
-    eq_split=item.split("=")
-    key = eq_split[0][1:].strip()
-    value = eq_split[1].strip()
-    ret[key]=value
-  return ret
-
-
-def parsedToUml(parsed, name, xscale, yscale):
- cog.outl("/\x2a*")
- cog.outl("@startuml")
- cog.outl("title "+name)
- if xscale is not None and yscale is not None:
-  cog.outl("scale {:d}*{:d}".format(xscale, yscale))
- elif xscale is not None:
-  cog.outl("scale {:d} width".format(xscale))
- elif yscale is not None:
-  cog.outl("scale {:d} height".format(yscale))
- for key in parsed.keys():
-  if key != "CH_TOTAL":
-    name = key[key.find("CH_") + 3:]
-    cog.outl("binary \""+name+"\" as " +key)
- totalms = int(parsed["CH_TOTAL"][0]["ms"])
- for key in parsed.keys():
-  if key != "CH_TOTAL":
-    l = parsed[key]
-    for n,element in enumerate(l):
-      cog.outl("@"+element["ms"])
-      if element["Mode"] == "OUT_MODE_ON":
-        cog.outl(key + " is high")
-      if element["Mode"] == "OUT_MODE_OFF":
-        cog.outl(key + " is low")
-      if element["Mode"] == "OUT_MODE_PULSED":
-        start = int(element["ms"])
-        on = int(element["On"])
-        off = int(element["Off"])
-        if n<len(l) - 1: #not the last element
-         to = int(l[n + 1]["ms"])
-        else:
-          to = totalms
-        ts = int(element["ms"])
-        cog.outl(key + " is high")
-        while True:
-          ts += on
-          if ts >= to:
-            break
-          cog.outl("@+" + str(on))
-          cog.outl(key + " is low")
-          ts += off
-          if ts >= to:
-            break
-          cog.outl("@+" + str(off))
-          cog.outl(key + " is high")
-
- cog.outl("@enduml")
- cog.outl("\x2a/")
-
-
-def convertToUml(instanceNo,xscale = None, yscale = None):
-  parsed = {}
-  with open(cog.inFile) as f:
-    instance = 0
-    for line in f:
-      split = line.split()
-      if checkCyclogram(split):
-          instance += 1
-      if (instance >= instanceNo + 1):
-        break
-    if instance != instanceNo + 1:
-      cog.error("No Cycloram line found")
-    else:
-      name = split[2]
-      name = name[0:name.find("[")]
-      for line in f:
-        if line.find(";") != -1 :
-          break
-        dict = line2dict(line)
-        if "Channel" in dict.keys():
-          dictkey = dict["Channel"]
-          dict.pop("Channel")
-          if dictkey not in parsed.keys():
-            parsed[dictkey] = []
-          parsed[dictkey].append(dict)
-      #cog.msg(str(parsed))
-      parsedToUml(parsed, name, xscale, yscale)
-
-  f.close()
-
-convertToUml(0)
+convert_to_uml(0)
 
 ]]]*/
 /**
@@ -137,7 +43,7 @@ const Channel_Step_t Cyclogram_Cold[]=
 };
 
 /*[[[cog
-convertToUml(1,3000,500)
+convert_to_uml(1,3000,500)
 ]]]*/
 /**
 @startuml
